combine secrets-manager with secret for CLI
