+++ src/zenml/cli/__init__.py
@@ -818,7 +818,6 @@ from zenml.cli.example import *  # noqa
 from zenml.cli.feature import *  # noqa
 from zenml.cli.integration import *  # noqa
 from zenml.cli.pipeline import *  # noqa
-from zenml.cli.secret import *  # noqa
 from zenml.cli.served_models import *  # noqa
 from zenml.cli.server import *  # noqa
 from zenml.cli.stack import *  # noqa
